Adding option for compiling different cuda versions
