+++ create-all-binaries.sh
@@ -3,7 +3,7 @@
 set -euxo pipefail
 
 declare -a platforms=("windows-x86_64" "macosx-x86_64" "linux-x86_64")
-declare -a chips=("cpu" "gpu")
+declare -a chips=("gpu")
 declare -a spins=("minimal" "pmml" "python" "all")
 
 mkdir -p builds
@@ -21,10 +21,22 @@ do
         continue
       fi
 
-      echo "Compiling for $platform | $chip | $spin"
+      if [[ "${chip}" == 'gpu' ]]; then
+        declare -a cuda_versions=("10.1" "10.2")
 
-      python build_jar.py --os "${platform}" --chip "${chip}" --spin "${spin}" \
-          --target builds/konduit-serving-uberjar-"${PROJECT_VERSION}"-"${spin}"-"${platform}"-"${chip}".jar
+        for cuda_version in "${cuda_versions[@]}"
+        do
+          echo "Compiling for ${platform} | gpu | ${cuda_version} | ${spin}"
+
+          python build_jar.py -p "${platform}" -c "${chip}" -cv ${cuda_version} -s "${spin}" \
+              --target builds/konduit-serving-uberjar-"${PROJECT_VERSION}"-"${spin}"-"${platform}"-"${chip}"-${cuda_version}.jar
+        done
+      else
+        echo "Compiling for ${platform} | ${chip} | ${spin}"
+
+        python build_jar.py -p "${platform}" -c "${chip}" -s "${spin}" \
+            --target builds/konduit-serving-uberjar-"${PROJECT_VERSION}"-"${spin}"-"${platform}"-"${chip}".jar
+      fi
     done
   done
 done
